requirements.txt
# 第三方库依赖
six==1.17.0
requests==2.31.0
pyyaml==6.0.1
curl_cffi==0.9.0
fake_useragent==1.5.1
urllib3==1.25.11